Add: error messages that indicate twidlr in use
A concern voiced by users was that they may forget whether twidlr is being used or not, and which functions are available etc. Without writing new model names, this commit adds specific error messages that alert the user if they provide arguments that are not compatible with the twidlr function. The error messages explicitly mention that the function comes from twidlr